README.md
This project was completed as part of the FreeCodeCamp curriculum. It involves building a Convolutional Neural Network (CNN) with TensorFlow 2.0 and Keras to classify images of cats and dogs. The model is trained on a labeled dataset and achieves over 63% accuracy, utilizing data augmentation techniques to enhance performance and prevent overfitting.